requirements.txt
# ================================================================
# ClapLog - Film Production Tracker
# Requirements File
# ================================================================

# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1

# Database
mysqlclient==2.2.1

# WebSocket & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Streamlit
streamlit==1.30.0
streamlit-option-menu==0.3.6

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Generation
reportlab==4.0.7

# Image Processing
Pillow==10.2.0

# API & HTTP
requests==2.31.0
httpx==0.26.0

# Environment
python-decouple==3.8

# Security
argon2-cffi==23.1.0

# Utilities
pytz==2023.3.post1

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
